requirements.txt
fastapi
uvicorn
requests
lxml
PyYAML
